requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
numpy==1.23.3
openai==0.27.0
pandas==1.4.3
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.1
PyYAML==6.0
requests==2.32.3
six==1.16.0
tqdm==4.64.0
urllib3==2.2.2
yarl==1.9.4